+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ByteriX, 2017. All right reserved.
 ## 2.3.3 (04.05.2017)
 ##### Improvements
 * change public/open accessible on any classes
+* added param willSelect: Bool to all check functions.
 
 ## 2.3.2 (03.05.2017)
 ##### Improvements